api-segment/project/requirements.txt
django==1.9
djangorestframework==3.3.3

six>=1.4.1
# unicodecsv
unicodecsv==0.14.1

# mongoengine
# pymongo
mongoengine==0.9
pymongo==2.7

psycopg2

django-rest-framework-mongoengine
# djangorestframework-csv
djangorestframework-csv==2.0.0

# test

appnope==0.1.0
certifi==2017.11.5
decorator==4.1.2

ipdb==0.10.3
#ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
simplegeneric==0.8.1
traitlets==4.3.2
wcwidth==0.1.7
